% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiocDockerManager.R
\docType{package}
\name{BiocDockerManager}
\alias{BiocDockerManager}
\title{BiocDockerManager: Manage Bioconductor docker images.}
\description{
A convenient package to install images, update images
 and find which Bioconductor based docker images are 
 available.
}
\section{BiocDockerManager functions}{


available()

installed()

install()

version()

valid()

use_dockerfile()
}

\author{
maintainer@bioconductor.org
}
