% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{.create_payload}
\alias{.create_payload}
\title{Fill in the template with a payload.}
\usage{
.create_payload(version = c("devel", "release"))
}
\arguments{
\item{`character(1)`}{version for Bioconductor is 'devel' or 'release'}
}
\description{
Fill in the template with a payload.
}
\details{
`create_payload()` fills in the template with
    the appropriate payload. The payload can be either
    for release or devel.
}
