% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_api.R
\name{.docker_do}
\alias{.docker_do}
\title{Docker command line function to run with arguments}
\usage{
.docker_do(args)
}
\arguments{
\item{args}{arguments to docker command in a character vector}
}
\description{
Docker command line function to run with arguments
}
\keyword{internal}
