% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerhub_api.R
\name{.docker_get}
\alias{.docker_get}
\title{Build path for Dockerhub API query}
\usage{
.docker_get(
  path,
  api = "https://hub.docker.com/v2",
  path_root = "/repositories/"
)
}
\description{
Build path for Dockerhub API query
}
