% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerhub_api.R
\name{.docker_get_query}
\alias{.docker_get_query}
\title{Dockerhub API get query}
\usage{
.docker_get_query(query)
}
\arguments{
\item{`character(1)`, }{query for the api.}
}
\value{
reponse of the query to the dockerhub API.
}
\description{
Dockerhub API get query
}
\keyword{internal}
