% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerhub_api.R
\name{.docker_image_description}
\alias{.docker_image_description}
\title{Get docker image description}
\usage{
.docker_image_description(repository)
}
\arguments{
\item{repository}{`character(1)` repository name}
}
\description{
Get docker image description
}
\keyword{internal}
