% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerhub_api.R
\name{.docker_image_list}
\alias{.docker_image_list}
\title{Get the list of images under an organization account.}
\usage{
.docker_image_list(organization = character(1))
}
\arguments{
\item{`character(1)`}{name of organization or "username".}
}
\value{
`character(1)` vector of images hosted by the organization
    on dockerhub.
}
\description{
Get the list of images under an organization account.
}
\details{
Get the list of images under and organization or
    username. Important to note that organization/username is
    different from the image name on dockerhub.
}
\examples{
.docker_image_list("bioconductor")

}
\keyword{internal}
