% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerhub_api.R
\name{.docker_image_pull_count}
\alias{.docker_image_pull_count}
\title{Get Docker image pull count (number of times downloaded)}
\usage{
.docker_image_pull_count(image = character(1))
}
\arguments{
\item{`character(1)`}{docker image name with organization}
}
\value{
`numeric(1)` number showing how many times the image has
    been downloaded
}
\description{
Get Docker image pull count (number of times downloaded)
}
\examples{

\dontrun{
.docker_image_pull_count("bioconductor/bioconductor_docker")
}

}
\keyword{internal}
