% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_api.R
\name{.docker_inspect_digest}
\alias{.docker_inspect_digest}
\title{Inspect docker image digest value}
\usage{
.docker_inspect_digest(repository, tag)
}
\arguments{
\item{repository}{`character(1)`, repository name of the image.}

\item{tag}{`character(1)`, tag for the image.}
}
\description{
Inspect docker image digest value
}
\keyword{internal}
