% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_api.R
\name{.docker_pull}
\alias{.docker_pull}
\title{Pull a docker image, see 'install()'}
\usage{
.docker_pull(name, tag = "latest", quiet = FALSE, all_tags = FALSE)
}
\arguments{
\item{name}{`character(1)`, name of the image.}

\item{tag}{`character(1)`, name of the tag for the image. 'latest' is 
default.}

\item{quiet}{`logical(1)`, status of download is not displayed if TRUE.}

\item{all_tags}{`logical(1)`, pull all the tags of the image name.}
}
\value{
exit status of `docker_pull()` command.
}
\description{
Pull a docker image, see 'install()'
}
\examples{

\dontrun{
.docker_pull("bioconductor/bioconductor_docker")

.docker_pull("bioconductor/bioconductor_docker", tag = "devel")

.docker_pull("bioconductor/bioconductor_docker", all_tags=TRUE)

.docker_pull("bioconductor/bioconductor_docker", quiet=TRUE)
}
}
\keyword{internal}
