% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{.dockerfile_template}
\alias{.dockerfile_template}
\title{Get dockerfile template}
\usage{
.dockerfile_template()
}
\description{
Get dockerfile template
}
\details{
Get the Dockerfile template which is annotated 
    with whisker templating tags.
}
