% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_api.R
\name{.FUN}
\alias{.FUN}
\title{Internal function to run commands using system2}
\usage{
.FUN(cmd, args, stdout = FALSE)
}
\arguments{
\item{cmd}{command line tool}

\item{args}{arguments for command line tool in a character vector}
}
\description{
This function has a tryCatch block to show the error message
when the wrong
}
\keyword{internal}
