% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{.get_rocker_rstudio_version}
\alias{.get_rocker_rstudio_version}
\title{Get rocker/rstudio image version}
\usage{
.get_rocker_rstudio_version(version = c("devel", "release"))
}
\arguments{
\item{version}{`character(1)` "release" or "devel" version.}
}
\value{
`character(1)` version number or name (devel).
}
\description{
Get rocker/rstudio image version
}
\details{
`get_rocker_rstudio_version()` gives the devel and release
    R version number of the latest release and devel images. These
    versions are taken from "inst/extdata/config".
}
