% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{help}
\alias{help}
\title{Help function to direct brower to Bioconductor dockerhub}
\usage{
help(repository = "bioconductor/bioconductor_docker")
}
\arguments{
\item{repository}{`character(1)`, repository name of image. Default
image is the main 'bioconductor/bioconductor_docker' image.}
}
\value{
Open a browser tab with docker repository
}
\description{
Help function to direct brower to Bioconductor dockerhub
}
\examples{
do_it <- BiocDockerManager:::.is_docker_installed()

if(do_it) {
    BiocDockerManager::help()
}
}
