% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{install}
\alias{install}
\title{Install a docker image on your local machine}
\usage{
install(repository, tag, quiet = FALSE, all_tags = FALSE)
}
\arguments{
\item{repository}{`character(1)`, repository name of the docker image.}

\item{tag}{`character(1)`, tag of the docker image.}

\item{quiet}{`logical(1)`, if TRUE suppress verbose output
generated from the download.}

\item{all_tags}{`logical(1)`, pull all the tags of the image}
}
\value{
invisible
}
\description{
Install a docker image on your local machine
}
\details{
The function works similar to the 'docker pull'
    command. It downloads a docker image from Dockerhub on to the
    local machine, in a place which the docker engine knows about,
    building your local registry of docker images.
}
\examples{

do_it <- BiocDockerManager:::.is_docker_installed()

if (do_it) {
    BiocDockerManager::install(
        repository = "bioconductor/bioconductor_docker",
        tag = "latest"
    )
}

}
