% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{installed}
\alias{installed}
\title{Get installed docker images}
\usage{
installed(name)
}
\arguments{
\item{name}{`character(1)`, name of image; if not given
all images will be shown.}
}
\value{
stdout of docker images on your local machine.
}
\description{
Get installed docker images
}
\examples{
\dontrun{
   BiocDockerManager::installed()

   BiocDockerManager::installed(name = "bioconductor/bioconductor_docker")
}

}
