% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{render_dockerfile}
\alias{render_dockerfile}
\title{Render the template with the}
\usage{
render_dockerfile(version = c("devel", "release"), path)
}
\arguments{
\item{`character(1)`}{path to where the rendered file is written.}
}
\description{
Render the template with the
}
\details{
`render_dockerfile()` will render the 
    dockerfile from the template with the appropriate
    information for release or devel.
}
