% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{use_dockerfile}
\alias{use_dockerfile}
\title{Use dockerfile template}
\usage{
use_dockerfile(name, path)
}
\description{
Use dockerfile template
}
\examples{

\dontrun{
    
    use_dockerfile("custom_image", tempdir())

}
}
