% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerhub_api.R
\name{.docker_get}
\alias{.docker_get}
\title{Build query for Dockerhub API}
\usage{
.docker_get(
  path,
  api = "https://hub.docker.com/v2",
  path_root = "/repositories/"
)
}
\arguments{
\item{path}{`character(1)` query substring}

\item{api}{`character(1)`, dockerhub api url}

\item{path_root}{`character(1)` query substring}
}
\description{
Build query for Dockerhub API
}
\keyword{internal}
