% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_api.R
\name{.docker_images}
\alias{.docker_images}
\title{List installed images/available on local machine.}
\usage{
.docker_images(repository)
}
\arguments{
\item{repository}{`character(1)`, repository name of the image.}

\item{quiet}{`logical(1)`, output shows only IMAGE ID's if TRUE.}
}
\value{
tibble with the docker images
}
\description{
List installed images/available on local machine.
}
\examples{

\dontrun{
    .docker_images()

    .docker_images("bioconductor/bioconductor_docker")

    .docker_images("bioconductor/bioconductor_docker", quiet = TRUE)
}
}
\keyword{internal}
