% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_api.R
\name{.docker_inspect_label}
\alias{.docker_inspect_label}
\title{Inspect docker image by label}
\usage{
.docker_inspect_label(
  repository,
  tag,
  label = c("name", "description", "version", "url", "maintainer", "license", "vendor")
)
}
\arguments{
\item{repository}{`character(1)`, repository name of the image.}

\item{tag}{`character(1)`, tag for the image.}

\item{label}{`character(1)` label repository from the list - "repository",
"description", "version", "url", "maintainer",
"license","vendor"}
}
\description{
Inspect docker image by label
}
\keyword{internal}
