% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{installed}
\alias{installed}
\title{Get installed docker images}
\usage{
installed(name = "", ...)
}
\arguments{
\item{name}{`character(1)`, name of image; if not given
all images will be shown.}
}
\description{
Get installed docker images
}
