% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{installed}
\alias{installed}
\title{Get installed docker images}
\usage{
installed(repository)
}
\arguments{
\item{repository}{`character(1)`, repository name of image; if not
given all images will be shown.}
}
\value{
stdout of docker images on your local machine.
}
\description{
Get installed docker images
}
\examples{

BiocDockerManager::installed()

BiocDockerManager::installed(
    repository = "bioconductor/bioconductor_docker"
)

}
