% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dockerfile.R
\name{use_dockerfile}
\alias{use_dockerfile}
\title{Use dockerfile template}
\usage{
use_dockerfile(
  name,
  person,
  description,
  license = "Artistic-2.0",
  path = tempdir()
)
}
\arguments{
\item{name}{`character(1)`, name of the image.}

\item{person}{`person(1)`, class person which holds information
about persons like name, email address, and maintainer/author
information.}

\item{path}{`character(1)`, parent path where folder repository
should be created.}
}
\description{
Use dockerfile template
}
\examples{

\dontrun{

    use_dockerfile("custom_image", tempdir())

}
}
