% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{valid}
\alias{valid}
\title{Check if all images available are valid}
\usage{
valid(repository = "bioconductor/bioconductor_docker", tag)
}
\arguments{
\item{repository}{`character(1)`, repository name of the docker image.}

\item{tag}{`character(1)`, tag of the docker image.}
}
\value{
tibble with the repository and tag of image which needs to be
    updated.
}
\description{
Check if all images available are valid
}
\details{
Check if the image is valid, i.e to see if the image is up
    to date with the image hosted by bioconductor on the Dockerhub
    organization page.
}
\examples{
\dontrun{

BiocDockerManager::valid()

BiocDockerManager::valid("bioconductor/bioconductor_docker",
                         tag = "devel")
}

}
