% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{version}
\alias{version}
\title{Get version of bioconductor docker image}
\usage{
version(repository = "bioconductor/bioconductor_docker", tag)
}
\arguments{
\item{repository}{`character(1)`, repository of the docker image.}

\item{tag}{`character(1)`, tag of the docker image.}
}
\value{
`character` vector representing the version number.
}
\description{
Get version of bioconductor docker image
}
\details{
The version of the images provided by Bioconductor are
    unique. They are represented by 'x.y.z' where, the 'x.y'
    represent the version of Bioconductor and the '.z' represents
    the version of the Dockerfile used to build the Docker
    image. This is especially useful in terms of reproducibility
    and tracking changes when using the Docker images provided by
    Bioconductor.
}
\examples{
BiocDockerManager::version(
    "bioconductor/bioconductor_docker",
    tag = "latest")

BiocDockerManager::version(
    "bioconductor/bioconductor_docker",
    tag = "devel")

}
